feat: Remove dependency on just
Since its just (pun intended) a bit switch-case statement anyways,
I rewrote it as a minimal bash script instead. This requires
re-running the bootstrap.sh in order to symlink the ~/.bin folder
to make the just command available.

Also moved the Google stuff out of the regular Silverblue install,
as I only need it on certain installs not all. Now available as a
just command instead.